Am facut logica de logare si am adaugat contactul cu magazinul

+++ src/main/java/com/application/mamuca/repository/UserRepository.java
@@ -2,7 +2,10 @@ package com.application.mamuca.repository;
 
 import com.application.mamuca.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.security.core.userdetails.UserDetails;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
+
     User findByEmail(String email);
+
 }
